requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
peewee==3.17.0
PyMySQL==1.1.0
mysql-connector-python==8.2.0

# Redis
redis==5.0.1
hiredis==2.3.2

# AI/ML
langchain==0.0.350
openai==1.3.0
anthropic==0.7.8
requests==2.31.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# JSON and Data Processing
ujson==5.8.0
orjson==3.9.10

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation and Parsing
email-validator==2.1.0
python-dateutil==2.8.2

# Logging
loguru==0.7.2
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0
py-cpuinfo==9.0.0
psutil==5.9.6

# Environment
python-dotenv==1.0.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality (Development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Development Tools
watchfiles==0.21.0